requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==73.0.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
